% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_shinyFiles.R
\name{dirChooserServer}
\alias{dirChooserServer}
\title{A formatted directory input using shinyFiles
The format is based on AdminLTE v2 (shinydashboard)}
\usage{
dirChooserServer(id, roots, allowDirCreate = TRUE)
}
\arguments{
\item{id}{Module id.}

\item{roots}{A named vector of absolute filepaths or a function returning a named vector of absolute filepaths (the latter is useful if the volumes should adapt to changes in the filesystem).}

\item{allowDirCreate}{Logical that indicates if creating new directories by the user is allowed.}
}
\description{
A formatted directory input using shinyFiles
The format is based on AdminLTE v2 (shinydashboard)
}
