% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_shinyFiles.R
\name{dirChooserUI}
\alias{dirChooserUI}
\title{A formatted directory input using shinyFiles
The format is based on AdminLTE v2 (shinydashboard)}
\usage{
dirChooserUI(id, label, title)
}
\arguments{
\item{id}{Module id.}

\item{label}{The button label.}

\item{title}{The browser window title.}
}
\description{
A formatted directory input using shinyFiles
The format is based on AdminLTE v2 (shinydashboard)
}
