% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_shinyFiles.R
\name{dirChooserWithResetServer}
\alias{dirChooserWithResetServer}
\title{A formatted directory input with the UI rendered from the server to enable resetting.
The format is based on AdminLTE v2 (shinydashboard)}
\usage{
dirChooserWithResetServer(id, roots, label, title, reset)
}
\arguments{
\item{id}{Module id.
#' @param roots     A named vector of absolute filepaths or a function returning a named vector of absolute filepaths (the latter is useful if the volumes should adapt to changes in the filesystem).}

\item{label}{The button label.}

\item{title}{The browser window title.}

\item{reset}{A reactive value to trigger the re-rendering of the UI.}
}
\description{
A formatted directory input with the UI rendered from the server to enable resetting.
The format is based on AdminLTE v2 (shinydashboard)
}
