% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_shinyFiles.R
\name{dirChooserWithResetUI}
\alias{dirChooserWithResetUI}
\title{A formatted directory input with the UI rendered from the server function to enable resetting.
The format is based on AdminLTE v2 (shinydashboard)}
\usage{
dirChooserWithResetUI(id)
}
\arguments{
\item{id}{Module id.}
}
\description{
A formatted directory input with the UI rendered from the server function to enable resetting.
The format is based on AdminLTE v2 (shinydashboard)
}
