% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_shinyFiles.R
\name{fileChooserServer}
\alias{fileChooserServer}
\title{A formatted file input using shinyFiles
The format is based on AdminLTE v2 (shinydashboard)}
\usage{
fileChooserServer(id, roots)
}
\arguments{
\item{id}{Module id.}

\item{roots}{A named vector of absolute filepaths or a function returning a named vector of absolute filepaths (the latter is useful if the volumes should adapt to changes in the filesystem).}
}
\description{
A formatted file input using shinyFiles
The format is based on AdminLTE v2 (shinydashboard)
}
