% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_shinyFiles.R
\name{fileChooserWithResetServer}
\alias{fileChooserWithResetServer}
\title{A formatted directory input using shinyFils with the UI rendered from the server to enable resetting.
The format is based on AdminLTE v2 (shinydashboard)}
\usage{
fileChooserWithResetServer(
  id,
  roots,
  reset,
  label,
  title,
  multiple = TRUE,
  filetypes = NULL,
  pattern = NULL
)
}
\arguments{
\item{id}{Module id.}

\item{roots}{A named vector of absolute filepaths or a function returning a named vector of absolute filepaths (the latter is useful if the volumes should adapt to changes in the filesystem).}

\item{reset}{A reactive value to trigger the re-rendering of the UI.}

\item{label}{The button label.}

\item{title}{The browser window title.}

\item{multiple}{A logical indicating whether or not it should be possible to select multiple files}

\item{filetypes}{A character vector of file extensions (without dot in front i.e. 'txt' not '.txt') to include in the output. Use the empty string to include files with no extension. If not set all file types will be included}

\item{pattern}{A regular expression used to select files to show. See base::grepl() for additional discussion on how to construct a regular expression (e.g., "log.*\\.txt")}
}
\description{
A formatted directory input using shinyFils with the UI rendered from the server to enable resetting.
The format is based on AdminLTE v2 (shinydashboard)
}
