% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_shinyFiles.R
\name{fileChooserWithResetUI}
\alias{fileChooserWithResetUI}
\title{A formatted directory input using shinyFils with the UI rendered from the server to enable resetting.
The format is based on AdminLTE v2 (shinydashboard)}
\usage{
fileChooserWithResetUI(id)
}
\arguments{
\item{id}{Module id.}
}
\description{
A formatted directory input using shinyFils with the UI rendered from the server to enable resetting.
The format is based on AdminLTE v2 (shinydashboard)
}
